requirements.txt
bcrypt==4.0.1
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.7
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
